% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessSpatialBias_modified.R
\name{assessSpatialBias_modified}
\alias{assessSpatialBias_modified}
\title{\code{assessSpatialBias}}
\usage{
assessSpatialBias_modified(
  dat,
  species,
  x,
  y,
  year,
  spatialUncertainty,
  identifier,
  periods,
  mask,
  nSamps = 50,
  degrade = TRUE,
  maxSpatUncertainty = NULL,
  crs = "BNG"
)
}
\arguments{
\item{dat}{string. A data.frame containing columns for species name, x coordinates, y coordinates, spatialUncertainty, year and an identifier (used to group the data - heuristic will be calculated for each group).}

\item{species}{Character string. column name in dat giving species names.}

\item{x}{string. Column name in dat giving x coordinates.}

\item{y}{string. Column name in dat giving y coordinates.}

\item{year}{string. Column name in dat giving years.}

\item{spatialUncertainty}{String. Column name in dat giving uncertainty associated with x and y. Any units are permitted.}

\item{identifier}{String. Column name in dat giving record "identifiers". Identifiers are used to group the data; heuristics will be calculated separately for each group.}

\item{periods}{Numeric. A list of time periods. For example, for two periods, the first spanning 1950 to 1990, and the second 1991 to 2019: periods = list(1950:1990, 1991:2019).}

\item{mask}{String. A raster object used to indicate the study region over which the random distribution should be generated. In most cases this will be a single raster layer. However, where the identifier 
     field is used to subset the data spatially (e.g. splitting it by country, continent, etc.) then mask should be a raster stack with n layers, where n is equal to the number of levels in the identifier field. 
     mask may have fewer layers than the number of levels in identifier field when identifier refers to spatial subsets, but any rows in dat with an identifier for which there is no layer 
     in mask will be dropped. If nlayers(mask) > 1, i.e. if the identifier field is used to subset the data spatially, then names(mask) should match the names in the identifier field.
Must be NA where points are not to be generated, and numeric where they may be generated. For example, this could be a map of worldclim climate data, cropped to the study region.}

\item{nSamps}{Logical. How many iterations of random samples to use for comparison of empirical NN index with random NN index.}

\item{degrade}{Logical. Whether or not to remove duplicated coordinates from the data. Coordinates are not considered to be duplicated if they are from
different \code{periods}.}

\item{maxSpatUncertainty}{Numeric. Maximum permitted spatial uncertainty. All records more uncertain than this value will be dropped. Units must match the units in your data.}

\item{crs}{Either "BNG" or "WGS84", specifying the CRS of the x/y input.}
}
\value{
A list with two elements: a ggplot2 object and the data underpinning the plot.
}
\description{
This function calculates how far the data deviates from a random distribution in geographic space. It calculates
a nearest neighbour index, defined as the mean of the nearest neighbour distances of the emprical data divided by
the mean of the neares neighbour distances of a random sample. The user can choose how many random samples to average
over, which should be larger where there are fewer data. This is because the random samples are generated in equal number to
the data. Where the number of samples is > 1, the estimates come with a measures of uncertainty (5th and 95th percentiles). 
The index is calculated for each of n user-specified time periods.
}
\seealso{
\code{\link{assessSpatialCov}} which maps your data in geographical space.
}
