% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_file.R, R/countriesLow copy.R
\docType{data}
\name{aux_file}
\alias{aux_file}
\title{environmental variables in the UK}
\usage{
aux_file

aux_file
}
\description{
environmental variables in the UK

environmental variables in the UK
}
