% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_descriptions.R
\docType{data}
\name{variable_descriptions}
\alias{variable_descriptions}
\title{environmental variables description in the UK}
\usage{
variable_descriptions
}
\description{
environmental variables description in the UK
}
