% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessBias1D_modified.r
\name{assessBias1D_modified}
\alias{assessBias1D_modified}
\title{\code{assessBias1D_modified}}
\usage{
assessBias1D_modified(pop, R, x, weights = NULL, breaks, RNames, WNames = NULL)
}
\arguments{
\item{pop}{string. A data.frame with one row for every spatial unit in the population (i.e. the geographic domain) and several columns. At least one column must be a binary variable (1 or 0) taking the value 1 if
a spatial unit is in the sample and 0 otherwise. More than one column can contain information on sample inclusion if, say, there are multiple time-periods (see R).}

\item{R}{String. Name(s) of the column(s) in pop that indicate whether each unit was included in the sample. If length(R) >1, then several sets of outputs will be provided, one for each R.}

\item{x}{String. Name of the column in pop with the variable whose distributions will be assessed.}

\item{weights}{Numeric vector or list of numeric vectors. One or more sets of weights whose lengths equal the number of rows in pop. Each weight should should correspond to the matching row of pop. Weights can be calculated using e.g. poststratification,
superpopulation modelling, etc. (Boyd et al. 2023).}

\item{breaks}{Numeric. \code{assessBias1D_modified} calculates a relative frequency distribution of x in the sample and population, and breaks is the number of bins into which the distributions will be split.}

\item{RNames}{String or character vector. Names of sample inclusion variables R. One name per element of R. Used for plots and other outputs.
#' @param RNames String or character vector. Names of sample inclusion variables R. One name per element of R. Used for plots and other outputs.}

\item{WNames}{String or character vector. Name of each set of weights. One name per set (this will often be 1). Used for plots and other outputs.}
}
\value{
a list with two elements: 1) a plot of the relative frequency distributions and 2) the underlying data.
}
\description{
This function compares the distribution of some variable in the sample to its distribution in the population (i.e. the whole geographic domain). If a set of weights are provided, it also compares the weighted
distribution in the sample to the distribution in the population. Multiple sets of weights can be supplied, in which case the function will produce several plots datasets, each pertaining to one set. The idea 
is that a sample is representative, at least in terms of the focal variable, if the sample and population distributions are similar. If the weighted sample distribution is closer than the raw sample distribution 
to the populaiton distribution, then weighting was successful. Note that the data must be provided in a different format for this function than the others.
}
